+++ update/3610/preupdate.php
@@ -1,56 +1,60 @@
 <?php
 
-// QR-CODES
-$dir = PIMCORE_CONFIGURATION_DIRECTORY . "/qrcodes";
-
-$file = Pimcore\Config::locateConfigFile("qrcode.json");
-$json = \Pimcore\Db\JsonFileTable::get($file);
-$json->truncate();
-
-$files = scandir($dir);
-foreach ($files as $file) {
-    if(strpos($file, ".xml")) {
-        $name = str_replace(".xml", "", $file);
-        $thumbnail = \Pimcore\Model\Tool\Qrcode\Config::getByName($name);
-        $thumbnail = object2array($thumbnail);
-
-        $thumbnail["id"] = $thumbnail["name"];
-        unset($thumbnail["name"]);
-
-        $json->insertOrUpdate($thumbnail, $thumbnail["id"]);
-    }
-}
-
 // create legacy config folder
 $legacyFolder = PIMCORE_CONFIGURATION_DIRECTORY . "/LEGACY";
-if(!is_dir($legacyFolder)) {
+if (!is_dir($legacyFolder)) {
     mkdir($legacyFolder, 0777, true);
 }
 
-// move data
-rename($dir, $legacyFolder . "/qrcodes");
+// QR-CODES
+$dir = PIMCORE_CONFIGURATION_DIRECTORY . "/qrcodes";
+
+if(is_dir($dir)) {
+    $file = Pimcore\Config::locateConfigFile("qrcode.json");
+    $json = \Pimcore\Db\JsonFileTable::get($file);
+    $json->truncate();
+
+    $files = scandir($dir);
+    foreach ($files as $file) {
+        if (strpos($file, ".xml")) {
+            $name = str_replace(".xml", "", $file);
+            $thumbnail = \Pimcore\Model\Tool\Qrcode\Config::getByName($name);
+            $thumbnail = object2array($thumbnail);
+
+            $thumbnail["id"] = $thumbnail["name"];
+            unset($thumbnail["name"]);
+
+            $json->insertOrUpdate($thumbnail, $thumbnail["id"]);
+        }
+    }
+
+    // move data
+    rename($dir, $legacyFolder . "/qrcodes");
+}
 
 
 // SQL REPORTS
 $dir = PIMCORE_CONFIGURATION_DIRECTORY . "/sqlreport";
 
-$file = Pimcore\Config::locateConfigFile("custom-reports.json");
-$json = \Pimcore\Db\JsonFileTable::get($file);
-$json->truncate();
+if(is_dir($dir)) {
+    $file = Pimcore\Config::locateConfigFile("custom-reports.json");
+    $json = \Pimcore\Db\JsonFileTable::get($file);
+    $json->truncate();
 
-$files = scandir($dir);
-foreach ($files as $file) {
-    if(strpos($file, ".xml")) {
-        $name = str_replace(".xml", "", $file);
-        $thumbnail = \Pimcore\Model\Tool\CustomReport\Config::getByName($name);
-        $thumbnail = object2array($thumbnail);
+    $files = scandir($dir);
+    foreach ($files as $file) {
+        if (strpos($file, ".xml")) {
+            $name = str_replace(".xml", "", $file);
+            $thumbnail = \Pimcore\Model\Tool\CustomReport\Config::getByName($name);
+            $thumbnail = object2array($thumbnail);
 
-        $thumbnail["id"] = $thumbnail["name"];
-        unset($thumbnail["name"]);
+            $thumbnail["id"] = $thumbnail["name"];
+            unset($thumbnail["name"]);
 
-        $json->insertOrUpdate($thumbnail, $thumbnail["id"]);
+            $json->insertOrUpdate($thumbnail, $thumbnail["id"]);
+        }
     }
-}
 
-// move data
-rename($dir, $legacyFolder . "/sqlreport");
+    // move data
+    rename($dir, $legacyFolder . "/sqlreport");
+}
